+++ beacon-chain/rpc/prysm/v1alpha1/validator/proposer_test.go
@@ -721,7 +721,6 @@ func getProposerServer(db db.HeadAccessDatabase, headState state.BeaconState, he
 		FinalizationFetcher:   mockChainService,
 		ForkFetcher:           mockChainService,
 		ForkchoiceFetcher:     mockChainService,
-		MockEth1Votes:         true,
 		AttPool:               attestations.NewPool(),
 		SlashingsPool:         slashings.NewPool(),
 		ExitPool:              voluntaryexits.NewPool(),